Use a simpler approach to getting dependencies
